Update to 0.10.65 (#50)
* update to 0.10.61

* update to 0.10.61

* update to 0.10.61 part II

* update Dockerfile

* change workspacesPage to workspacePage

* update Dockerfile

* change workspacesPage to workspacePage

* update to 0.10.65
