+++ .env.development
@@ -1,2 +1,2 @@
 GEN3_COMMONS_NAME=bih
-NEXT_PUBLIC_GEN3_API=https://bihstaging.data-commons.org
+NEXT_PUBLIC_GEN3_API=https://imaging-hub.data-commons.org
