+++ .env.production
@@ -1,2 +1,1 @@
 GEN3_COMMONS_NAME=bih
-NEXT_PUBLIC_GEN3_API=
