+++ .nvmrc
@@ -1,1 +1,1 @@
-20
+v20.17.0
