+++ Dockerfile
@@ -31,22 +31,15 @@ RUN apt-get update && apt-get install -y --no-install-recommends \
     && npm install -g npm@10.5.2
 
 RUN  addgroup --system --gid 1001 nextjs && adduser --system --uid 1001 nextjs
-COPY ./package.json ./package-lock.json ./
-COPY ./package-lock.json ./
+COPY ./package.json ./package-lock.json ./next.config.js ./tsconfig.json ./tailwind.config.js ./postcss.config.js .env.production ./start.sh ./
 COPY ./src ./src
 COPY ./public ./public
 COPY ./config ./config
-COPY ./next.config.js ./
-COPY ./tsconfig.json ./
-COPY ./.env.development ./
-COPY ./.env.production ./
-COPY ./tailwind.config.js ./
-COPY ./postcss.config.js ./
-COPY ./start.sh ./
 RUN npm ci
 RUN npm install \
-    "@swc/core" \
-    "@napi-rs/magic-string"
+ "@swc/core" \
+ "@napi-rs/magic-string"
+
 RUN npm run build
 ENV PORT=3000
 CMD bash ./start.sh
