+++ config/icons/dataDictionary.json
@@ -1,0 +1,58 @@
+{
+	"prefix": "dataDictionary",
+	"lastModified": 1733715829,
+	"icons": {
+		"dictionary-icon-administrative": {
+			"body": "<g fill=\"currentColor\" fill-rule=\"evenodd\"><path d=\"M5.5 4.9h.72c.013.005.026.015.038.016.261.015.522.023.782.044.466.037.93.101 1.385.253.377.126.73.322 1.018.707.038.05.057.1.057.177q-.002 3.847-.001 7.696v.107c-.018-.007-.025-.008-.031-.012q-.118-.082-.234-.17a3.8 3.8 0 00-1.973-.803c-.51-.053-1.023-.07-1.534-.101-.07-.005-.142-.001-.22-.001V5.036c0-.046-.005-.09-.007-.136m8.28 0h.72q-.005.067-.007.135v7.63l-.004.136h-.077a13 13 0 00-1.784.118 4 4 0 00-1.863.79c-.084.064-.17.123-.265.191v-.107q0-3.854.002-7.707c0-.051.02-.115.045-.15q.222-.309.509-.484c.33-.205.68-.304 1.035-.375.506-.1 1.015-.14 1.526-.158.054-.002.108-.013.162-.019\"/><path d=\"M3.5 5.9h.983v8.114h.12c.61.012 1.22.01 1.828.039.717.033 1.434.105 2.121.374.316.123.618.295.925.446.01.004.016.018.023.027h-6zm8 9c.006-.01.011-.022.019-.028.5-.36 1.027-.598 1.584-.702a10 10 0 011.686-.148l.799-.008h.093V5.9h.819v9z\"/></g>"
+		},
+		"dictionary-icon-analysis": {
+			"body": "<g fill=\"currentColor\" transform=\"rotate(-90 10 5)\"><rect width=\"4.5\" height=\"2.7\" rx=\".5\"/><rect width=\"7.2\" height=\"2.7\" y=\"3.6\" rx=\".5\"/><path d=\"M.5 7.2h8a.5.5 0 01.5.5v1.7a.5.5 0 01-.5.5h-8a.5.5 0 01-.5-.5V7.7a.5.5 0 01.5-.5\"/></g>",
+			"height": 21
+		},
+		"dictionary-icon-biospecimen": {
+			"body": "<path fill=\"currentColor\" fill-rule=\"evenodd\" d=\"M11.54 8.807c-.091-.111-.131-.15-.13-.29.005-.942.002-1.957.002-2.9q-.001-1.106.002-2.212c.001-.355-.048-.405-.429-.405H9.12c-.278 0-.362.08-.362.344V5.74c.005.798.004 2.431-.004 2.777a.44.44 0 01-.092.251 701 701 0 01-3.96 4.837c-.218.263-.272.501-.101.787.006.01.262.52.456.579q.195.059 9.447 0c.206 0 .628-.137.892-.437.107-.122.202-.715-.095-1.06-1.321-1.618-2.438-3.051-3.761-4.667\"/>"
+		},
+		"dictionary-icon-clinical": {
+			"body": "<path fill=\"currentColor\" d=\"M5.45 8.053h2.61V5.45c0-.359.29-.65.65-.65h2.596a.65.65 0 01.648.65v2.602h2.596c.36 0 .65.291.65.65v2.601a.65.65 0 01-.65.652h-2.596v2.595c0 .36-.29.65-.648.65H8.709a.65.65 0 01-.65-.65v-2.595H5.452a.65.65 0 01-.651-.652v-2.6a.65.65 0 01.65-.65\"/>"
+		},
+		"dictionary-icon-clinical-assessment": {
+			"body": "<defs><path id=\"svgID0\" d=\"M12.56 9.434c.66.714.551 1.963-.223 2.794-.441.445-1.555 1.18-2.539.294v.866c0 .326-.25.595-.552.595H.552c-.303 0-.552-.27-.552-.595V1.677c0-.327.249-.595.552-.595h1.934V.605c0-.326.249-.594.552-.594h3.727c.303 0 .552.268.552.594v.477h1.932c.302 0 .551.268.551.595v4.786zM3.59 1.17v1.01h2.621V1.17zM1.077 12.82v.004h7.617V11.34L6.32 8.782c-.083-.088-.138-.177-.138-.297l-.47-2.734c-.053-.603.435-.731.608-.654l2.374.47V2.272h-1.38v.503c-.026.296-.275.566-.551.566H3.038c-.303 0-.552-.27-.552-.595v-.474H1.077zm10.488-1.424c.387-.418.414-.923.223-1.132L8.477 6.7l-1.574-.327.303 1.723 3.282 3.538c.39.297.856-.03 1.077-.237\"/></defs><use fill=\"currentColor\" fill-rule=\"evenodd\" href=\"#svgID0\" transform=\"translate(5 3)\"/>"
+		},
+		"dictionary-icon-data-file": {
+			"body": "<g fill=\"currentColor\" fill-rule=\"evenodd\"><path d=\"M6 9.748V4.541c0-.406.192-.603.59-.604l4.483-.004c.148 0 .27.057.374.163l3.287 3.344q.157.156.158.381l-.004 7.158c0 .301-.134.501-.382.567a.6.6 0 01-.164.017H6.544c-.339 0-.544-.21-.544-.556z\"/><path fill-rule=\"nonzero\" d=\"M10.3 3.933h1V8h-1z\"/><path fill-rule=\"nonzero\" d=\"M15 7.5v1h-4.7v-1z\"/></g>"
+		},
+		"dictionary-icon-data-observations": {
+			"body": "<g fill=\"currentColor\"><path d=\"m12.58 8.599.3-.301a.11.11 0 00.028-.11.11.11 0 00-.084-.077l-1.298-.263a.11.11 0 00-.131.131l.262 1.299a.11.11 0 00.187.056l.3-.3.704.703a.11.11 0 00.157 0l.278-.278a.11.11 0 000-.157z\"/><path d=\"m16.138 14.024-.647-1.89a.46.46 0 00-.432-.311h-.099a.89.89 0 00.62-.848V4.893A.89.89 0 0014.695 4h-9.49a.89.89 0 00-.885.893v6.082a.89.89 0 00.62.848h-.098a.46.46 0 00-.432.31l-.648 1.891a1.15 1.15 0 00.148 1.035c.212.3.554.48.92.48h10.24c.366 0 .708-.18.92-.48s.267-.686.148-1.035M9.95 4.206a.225.225 0 010 .451.225.225 0 01-.223-.225c0-.125.1-.226.223-.226m-4.745.657 9.52.03v6.111l-9.55-.03zm6.466 9.8a.13.13 0 01-.104.057H8.334a.13.13 0 01-.105-.056.13.13 0 01-.012-.118l.312-.78a.415.415 0 01.384-.261h2.074c.169 0 .32.103.384.261l.312.78a.13.13 0 01-.012.118\"/></g>"
+		},
+		"dictionary-icon-default": {
+			"body": "<g fill=\"currentColor\" transform=\"translate(4 9.5)\"><rect width=\"5\" height=\"2\" rx=\".5\"/><path d=\"M7.5 0h4a.5.5 0 01.5.5v1a.5.5 0 01-.5.5h-4a.5.5 0 01-.5-.5v-1a.5.5 0 01.5-.5\"/></g>",
+			"height": 21
+		},
+		"dictionary-icon-experimental-methods": {
+			"body": "<path fill=\"currentColor\" d=\"m15.807 6.539-1.976 1.976-1.729-.617-.617-1.73 1.976-1.975a3.723 3.723 0 00-4.737 4.642 2 2 0 00-.156.137L4.51 13.031a1.74 1.74 0 102.46 2.46l4.059-4.06q.075-.073.137-.155a3.722 3.722 0 004.642-4.737\"/>"
+		},
+		"dictionary-icon-imaging": {
+			"body": "<path fill=\"currentColor\" d=\"M15.81 5H4.19c-.105 0-.19.09-.19.2v9.6c0 .11.085.2.19.2h11.62c.105 0 .19-.09.19-.2V5.2c0-.11-.085-.2-.19-.2M15 13l-2.69-2.784a.08.08 0 00-.111-.005l-1.866 1.642-2.385-2.935a.08.08 0 00-.126.003L5 12.703V6h10z\"/>"
+		},
+		"dictionary-icon-index-file": {
+			"body": "<path fill=\"currentColor\" fill-rule=\"evenodd\" d=\"M4.5 5.5h3a1 1 0 110 2h-3a1 1 0 110-2m1 3h3a1 1 0 010 2h-3a1 1 0 110-2m6 0h3a1 1 0 010 2h-3a1 1 0 010-2m-4.5 3h3a1 1 0 010 2H7a1 1 0 010-2m6 0h3a1 1 0 010 2h-3a1 1 0 010-2m-2.5-6h3a1 1 0 010 2h-3a1 1 0 110-2\"/>"
+		},
+		"dictionary-icon-medical-history": {
+			"body": "<path fill=\"currentColor\" d=\"M5.45 8.053h2.61V5.45c0-.359.29-.65.65-.65h2.596a.65.65 0 01.648.65v2.602h2.596c.36 0 .65.291.65.65v2.601a.65.65 0 01-.65.652h-2.596v2.595c0 .36-.29.65-.648.65H8.709a.65.65 0 01-.65-.65v-2.595H5.452a.65.65 0 01-.651-.652v-2.6a.65.65 0 01.65-.65\"/>"
+		},
+		"dictionary-icon-metadata": {
+			"body": "<g fill=\"currentColor\" transform=\"translate(4.5 6)\"><rect width=\"5\" height=\"2\" rx=\".5\"/><path d=\"M6.5 0h4a.5.5 0 01.5.5v1a.5.5 0 01-.5.5h-4a.5.5 0 01-.5-.5v-1a.5.5 0 01.5-.5\"/><rect width=\"5\" height=\"2\" y=\"3\" rx=\".5\"/><path d=\"M6.5 3h4a.5.5 0 01.5.5v1a.5.5 0 01-.5.5h-4a.5.5 0 01-.5-.5v-1a.5.5 0 01.5-.5\"/><rect width=\"5\" height=\"2\" y=\"6\" rx=\".5\"/><path d=\"M6.5 6h4a.5.5 0 01.5.5v1a.5.5 0 01-.5.5h-4a.5.5 0 01-.5-.5v-1a.5.5 0 01.5-.5\"/></g>"
+		},
+		"dictionary-icon-notation": {
+			"body": "<path fill=\"none\" stroke=\"currentColor\" stroke-linecap=\"square\" d=\"M2.5 10.208h10m4-4.708-4 4.708m4 5.292-4-5.292M7.25 16.5v-12\"/>"
+		},
+		"dictionary-icon-study-administration": {
+			"body": "<g fill=\"currentColor\" fill-rule=\"evenodd\" transform=\"translate(4 5.5)\"><path fill-rule=\"nonzero\" d=\"M10.8 0H1.2C.54 0 0 .45 0 1v9l2.4-2h8.4c.66 0 1.2-.45 1.2-1V1c0-.55-.54-1-1.2-1\"/><circle cx=\"3.4\" cy=\"4\" r=\"1\"/><circle cx=\"6.4\" cy=\"4\" r=\"1\"/><circle cx=\"9.4\" cy=\"4\" r=\"1\"/></g>"
+		},
+		"dictionary-icon-subject-characteristics": {
+			"body": "<path fill=\"currentColor\" d=\"M10 9.73c1.38 0 2.5-1.156 2.5-2.88s-.367-3.12-2.5-3.12-2.5 1.397-2.5 3.12S8.62 9.73 10 9.73m5.494 5.173c-.054-3.346-.499-4.3-3.9-4.903 0 0-.478.6-1.594.6S8.406 10 8.406 10c-3.364.597-3.836 1.536-3.898 4.794-.005.266-.007.28-.008.25v.351s.81 2.062 5.5 2.062 5.5-2.062 5.5-2.062v-.26c0 .02-.003-.018-.006-.232\"/>"
+		}
+	},
+	"width": 20,
+	"height": 20
+}
