+++ config/icons/workspace.json
@@ -1,0 +1,24 @@
+{
+	"prefix": "workspace",
+	"lastModified": 1733715829,
+	"icons": {
+		"galaxy": {
+			"body": "<defs><linearGradient id=\"svgID0\" x1=\"100%\" x2=\"0%\" y1=\"50%\" y2=\"50%\"><stop offset=\"0%\" stop-color=\"currentColor\"/><stop offset=\"100%\" stop-color=\"currentColor\"/></linearGradient></defs><g fill=\"none\" fill-rule=\"evenodd\"><rect width=\"47.362\" height=\"13.945\" fill=\"currentColor\" rx=\"2\"/><rect width=\"35.176\" height=\"13.945\" y=\"18.09\" fill=\"currentColor\" rx=\"2\"/><rect width=\"49.121\" height=\"13.945\" x=\"16.206\" y=\"36.055\" fill=\"url(#svgID0)\" rx=\"2\"/></g>",
+			"width": 66
+		},
+		"jupyter": {
+			"body": "<use fill=\"currentColor\" href=\"#svgID0\" transform=\"translate(-1640 -2453) translate(1640.54 2474.36)\"/><use fill=\"currentColor\" href=\"#svgID1\" transform=\"translate(-1640 -2453) translate(1645.68 2474.37)\"/><use fill=\"currentColor\" href=\"#svgID2\" transform=\"translate(-1640 -2453) translate(1653.39 2474.26)\"/><use fill=\"currentColor\" href=\"#svgID3\" transform=\"translate(-1640 -2453) translate(1660.43 2474.39)\"/><use fill=\"currentColor\" href=\"#svgID4\" transform=\"translate(-1640 -2453) translate(1667.55 2472.54)\"/><use fill=\"currentColor\" href=\"#svgID5\" transform=\"translate(-1640 -2453) translate(1672.47 2474.29)\"/><use fill=\"currentColor\" href=\"#svgID6\" transform=\"translate(-1640 -2453) translate(1679.98 2474.24)\"/><use fill=\"currentColor\" href=\"#svgID7\" transform=\"translate(-1640 -2453) translate(1673.48 2453.69)\"/><use fill=\"currentColor\" href=\"#svgID8\" transform=\"translate(-1640 -2453) translate(1643.21 2484.27)\"/><use fill=\"currentColor\" href=\"#svgID9\" transform=\"translate(-1640 -2453) translate(1643.21 2457.88)\"/><use fill=\"currentColor\" href=\"#svgIDa\" transform=\"translate(-1640 -2453) translate(1643.28 2496.09)\"/><use fill=\"currentColor\" href=\"#svgIDb\" transform=\"translate(-1640 -2453) translate(1641.87 2458.43)\"/><defs><path id=\"svgID0\" d=\"M1.745 5.475c0 1.558-.125 2.066-.445 2.44a1.94 1.94 0 01-1.3.498l.125.89a3.05 3.05 0 002.03-.738 3.56 3.56 0 00.783-2.671V0H1.745z\"/><path id=\"svgID1\" d=\"M5.502 4.763c0 .668 0 1.264.053 1.78H4.496l-.071-1.059A2.47 2.47 0 012.26 6.695C1.23 6.695 0 6.135 0 3.846V.045h1.193v3.56c0 1.238.383 2.066 1.46 2.066A1.665 1.665 0 004.336 3.99V0h1.193v4.727z\"/><path id=\"svgID2\" d=\"M.053 2.273c0-.828 0-1.505-.053-2.12h1.068l.054 1.114A2.58 2.58 0 013.454.002c1.585 0 2.778 1.327 2.778 3.303 0 2.333-1.433 3.49-2.982 3.49a2.3 2.3 0 01-2.021-1.023v3.56H.053zM1.23 4.009q.004.242.053.48a1.834 1.834 0 001.78 1.38c1.256 0 1.995-1.023 1.995-2.51 0-1.3-.695-2.413-1.95-2.413a2.05 2.05 0 00-1.878 1.95z\"/><path id=\"svgID3\" d=\"M1.318.018 2.75 3.855c.151.427.312.944.418 1.327.125-.392.259-.89.419-1.354l1.3-3.81h1.255l-1.78 4.63c-.89 2.225-1.434 3.374-2.253 4.068a3.24 3.24 0 01-1.46.766l-.294-.997a3.2 3.2 0 001.042-.58 3.56 3.56 0 001.006-1.317.9.9 0 00.098-.285 1 1 0 00-.08-.311L0 0h1.3z\"/><path id=\"svgID4\" d=\"M2.19 0v1.87H3.9v.89H2.19v3.508c0 .801.232 1.264.89 1.264q.353.006.695-.08l.053.89c-.34.118-.7.172-1.06.16a1.66 1.66 0 01-1.29-.498 2.4 2.4 0 01-.463-1.692V2.751H0v-.89h1.033V.276z\"/><path id=\"svgID5\" d=\"M1.177 3.579A2.092 2.092 0 003.43 5.831a4.35 4.35 0 001.78-.338l.205.89a5.3 5.3 0 01-2.181.401A3.027 3.027 0 01.01 3.508C.01 1.549 1.177 0 3.082 0 5.22 0 5.753 1.87 5.753 3.063q.018.275 0 .552H1.15zm3.49-.89A1.683 1.683 0 003.011.766a1.97 1.97 0 00-1.825 1.923z\"/><path id=\"svgID6\" d=\"M.053 2.192c0-.765 0-1.424-.053-2.03h1.068v1.274h.054A1.97 1.97 0 012.902.01a1.3 1.3 0 01.339 0v1.113a1.8 1.8 0 00-.41 0 1.665 1.665 0 00-1.593 1.513 3.3 3.3 0 00-.054.552v3.464H.01V2.2z\"/><path id=\"svgID7\" d=\"M6.03 2.836A3.018 3.018 0 112.889.005a2.98 2.98 0 013.143 2.83\"/><path id=\"svgID8\" d=\"M18.696 7.122C10.684 7.122 3.641 4.247 0 0a19.934 19.934 0 0037.393 0c-3.633 4.247-10.65 7.122-18.697 7.122\"/><path id=\"svgID9\" d=\"M18.696 5.897c8.013 0 15.055 2.876 18.696 7.123A19.934 19.934 0 000 13.02c3.641-4.256 10.648-7.123 18.696-7.123\"/><path id=\"svgIDa\" d=\"M7.596 3.567A3.802 3.802 0 113.634.005a3.766 3.766 0 013.962 3.562\"/><path id=\"svgIDb\" d=\"M2.25 4.38A2.19 2.19 0 114.379 2.1a2.217 2.217 0 01-2.127 2.28\"/></defs>",
+			"width": 44,
+			"height": 51
+		},
+		"rlogo": {
+			"body": "<defs><linearGradient id=\"svgID0\" x1=\"0\" x2=\"1\" y1=\"0\" y2=\"1\" gradientUnits=\"objectBoundingBox\" spreadMethod=\"pad\"><stop offset=\"0\" stop-color=\"currentColor\"/><stop offset=\"1\" stop-color=\"currentColor\"/></linearGradient><linearGradient id=\"svgID1\" x1=\"0\" x2=\"1\" y1=\"0\" y2=\"1\" gradientUnits=\"objectBoundingBox\" spreadMethod=\"pad\"><stop offset=\"0\" stop-color=\"currentColor\"/><stop offset=\"1\" stop-color=\"currentColor\"/></linearGradient></defs><g fill-rule=\"evenodd\"><path fill=\"url(#svgID0)\" d=\"M361.453 485.937C162.329 485.937.906 377.828.906 244.469.906 111.109 162.329 3 361.453 3S722 111.109 722 244.469c0 133.359-161.422 241.468-360.547 241.468m55.188-388.531c-151.352 0-274.047 73.908-274.047 165.078s122.695 165.078 274.047 165.078c151.351 0 263.046-50.529 263.046-165.078 0-114.513-111.695-165.078-263.046-165.078\" transform=\"scale(.1)\"/><path fill=\"url(#svgID1)\" d=\"M550 377s21.822 6.585 34.5 13c4.399 2.226 12.01 6.668 17.5 12.5 5.378 5.712 8 11.5 8 11.5l86 145-139 .062L492 437s-13.31-22.869-21.5-29.5c-6.832-5.531-9.745-7.5-16.5-7.5h-33.026L421 558.974l-123 .052V152.938h247S657.5 154.967 657.5 262 550 377 550 377m-53.5-135.976-74.463-.048-.037 69.05 74.5-.024s34.5-.107 34.5-35.125c0-35.722-34.5-33.853-34.5-33.853\" transform=\"scale(.1)\"/></g>",
+			"width": 75
+		},
+		"rstudio": {
+			"body": "<g fill=\"currentColor\"><circle cx=\"24.983\" cy=\"24.983\" r=\"24.983\"/><path d=\"M33.524 34.177h2.71v2.105h-4.137l-6.855-10.274h-3.653v8.17h3.572v2.104H16.17v-2.105h3.09V16.073l-3.09-.38v-1.991c1.17.266 2.186.451 3.46.451 1.92 0 3.88-.451 5.798-.451 3.726 0 7.186 1.693 7.186 5.83 0 3.202-1.92 5.234-4.895 6.097zM21.59 23.976l1.96.04c4.782.073 6.62-1.766 6.62-4.218 0-2.863-2.072-3.992-4.701-3.992-1.242 0-2.524.113-3.88.267z\"/></g>",
+			"width": 50
+		}
+	},
+	"height": 50
+}
