+++ package.json
@@ -21,8 +21,8 @@
     "@fontsource/montserrat": "^5.0.19",
     "@fontsource/poppins": "^5.0.15",
     "@fontsource/source-sans-pro": "^5.0.8",
-    "@gen3/core": "^0.10.50",
-    "@gen3/frontend": "^0.10.50",
+    "@gen3/core": "^0.10.64",
+    "@gen3/frontend": "^0.10.64",
     "@grafana/faro-react": "^1.9.1",
     "@grafana/faro-web-sdk": "^1.9.1",
     "@grafana/faro-web-tracing": "^1.9.1",
@@ -35,7 +35,9 @@
     "@mdx-js/loader": "^3.0.0",
     "@mdx-js/react": "^3.0.0",
     "@next/mdx": "^14.2.15",
-    "cookies-next": "^4.1.1",
+    "cookies-next": "^4.3.0",
+    "idb": "^8.0.0",
+    "jsonpath-plus": "^10.2.0",
     "mantine-react-table": "^2.0.0-beta.7",
     "react": "^18.2.0",
     "react-dom": "18.2.0",
@@ -43,7 +45,7 @@
   },
   "devDependencies": {
     "@axe-core/react": "^4.10.0",
-    "@gen3/toolsff": "^0.10.50",
+    "@gen3/toolsff": "^0.10.65",
     "@tailwindcss/aspect-ratio": "^0.4.2",
     "@tailwindcss/forms": "^0.5.3",
     "@tailwindcss/line-clamp": "^0.4.2",
