+++ src/styles/globals.css
@@ -2,10 +2,14 @@
 @import "tailwindcss/components";
 
 @import '@mantine/core/styles.css';
+@import '@mantine/notifications/styles.css';
 @import 'mantine-react-table/styles.css';
 
 @import "tailwindcss/utilities";
 @import '@gen3/frontend/dist/styles.css';
+
+@import 'graphiql/graphiql.css';
+@import '@graphiql/react/dist/style.css';
 
 @layer base {
   html {
